MODIFY: movie/url, views 수정
- 별점 CRUD 메소드 및 로직 수정
- 별점 CRUD 구현을 위한 url 수정

+++ movie/views.py
@@ -23,42 +23,46 @@ class CreateStarView(View):
 
         try:
             star = Star.objects.create(
-                user_id = request.user,
+                user_id  = request.user,
                 movie_id = data["movieId"],
-                point = data["starPoint"]
+                point    = data["starPoint"]
             )
 
-            feedback = [
-                {"message":"SUCCESS"},
-                {"starPoint":star.point}
-            ]
-            return JsonResponse(feedback, status = 201)
+            feedback = {
+                "message"   : "SUCCESS",
+                "starPoint" : star.point
+            }
+            return JsonResponse(feedback, status=201)
 
         except KeyError:
-            return JsonResponse({"message":"KEY_ERROR"}, status = 400)
+            return JsonResponse({"message": "KEY_ERROR"}, status=400)
 
 
 class ReadStarView(View):
 #    <-- login decorator -->
     def get(self, request):
-        data = request.GET.get('movieId')
+        movie = request.GET.get('movieId')
+
+        if 'movieId' not in request.GET:
+            return JsonResponse({"message": "KEY_ERROR"}, status=400)
 
         star = Star.objects.filter(
-            user_id = request.user,
-            movie_id = data["movieId"]
+            user_id  = request.user,
+            movie_id = int(movie)
         )
 
         if star.exists():
-            star = star.first()
+            star       = star.first()
             star_point = star.point
         else:
-            return JsonResponse({"message":"NOT_FOUND"}, status =404)
+            return JsonResponse({"message": "NOT_FOUND"}, status=404)
 
-        feedback = [
-            {"message":"SUCCESS"},
-            {"starPoint":star_point}
-        ]
-        return JsonResponse(feedback, status = 200)
+        feedback = {
+            "message"   : "SUCCESS",
+            "starPoint" : star_point
+        }
+
+        return JsonResponse(feedback, status=200)
 
 
 class UpdateStarView(View):
@@ -67,43 +71,46 @@ class UpdateStarView(View):
         data = json.loads(request.body)
 
         try:
-            star = star.objects.filter(
-                user_id = request.user,
+            star = Star.objects.filter(
+                user_id  = request.user,
                 movie_id = data["movieId"]
             )
 
             if star.exists():
-                star.update(point = data["starPoint"])
+                star        = star.first()
+                star.point  = data["starPoint"]
+                star.save()
                 update_star = star.point
             else:
-                return JsonResponse({"message":"NOT_FOUND"}, status = 404)
+                return JsonResponse({"message": "NOT_FOUND"}, status=404)
 
-            feedback = [
-                {"message":"SUCCESS"},
-                {"starPoint":update_star}
-            ]
-            return JsonResponse(feedback, status = 200)
+            feedback = {
+                "message"   : "SUCCESS",
+                "starPoint" : update_star
+            }
+            return JsonResponse(feedback, status=200)
 
         except KeyError:
-            return JsonResponse({"message":"KEY_ERROR"}, status = 400)
+            return JsonResponse({"message": "KEY_ERROR"}, status=400)
 
 
 class DeleteStarView(View):
 #    <-- login decorator -->
     def delete(self, request):
-        data = request.GET.get(movieId)
+        movie = request.GET.get('movieId')
 
-        star = star.objects.filter(
-            user_id = request.user,
-            movie_id = data["movieId"]
+        if 'movieId' not in request.GET:
+            return JsonResponse({"message": "KEY_ERROR"}, status=400)
+
+        star = Star.objects.filter(
+            user_id  = request.user,
+            movie_id = int(movie)
         )
 
         if star.exists():
             star.delete()
         else:
-            return JsonResponse({"message":"NOT_FOUND"}, status = 404)
+            return JsonResponse({"message": "NOT_FOUND"}, status=404)
 
-        feedback = [
-            {"message":"SUCCESS"}
-        ]
-        return JsonResponse (feedback, status = 200)
+        feedback = {"message": "SUCCESS"}
+        return JsonResponse (feedback, status=200)
